+++ exercises/1901040047/d08/mymodule/main.py
@@ -1,0 +1,11 @@
+import stats_word
+test=3.14159265
+try:
+    
+    stats_word.stats_text(test)
+except ValueError:
+	    print('Input is not str,you may check again!')
+	
+
+
+
